PR-4 | Refactor authentication handling for journal entry operations. (#4)
- Moved user authentication retrieval to service methods for journal entry operations to streamline code and reduce redundant userName parameters.
- Added exception handling to `deleteEntry` method.
- Enhanced the endpoint methods in `JournalEntryController` to use the authenticated user's context.

+++ src/main/java/com/learning/journalApplication/service/JournalEntryService.java
@@ -46,10 +46,21 @@ public class JournalEntryService {
     }
 
     @Transactional
-    public void deleteEntry(ObjectId id, String userName){
-        User user = _userService.findUserByUserName(userName);
-        user.getJournalEntries().removeIf(journalEntry -> journalEntry.getId().equals(id));
-        _userService.saveUser(user);
-        _journalEntryRepository.deleteById(id);
+    public boolean deleteEntry(ObjectId id, String userName){
+        boolean isRemoved = false;
+        try {
+            User user = _userService.findUserByUserName(userName);
+            isRemoved = user.getJournalEntries()
+                    .removeIf(journalEntry -> journalEntry.getId().equals(id));
+            if (isRemoved) {
+                _userService.saveUser(user);
+                _journalEntryRepository.deleteById(id);
+            }
+        }
+        catch (Exception exception){
+            System.out.println(exception);
+            throw new RuntimeException("An error occurred while deleting the entry.", exception);
+        }
+        return isRemoved;
     }
 }
